requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.3.0
